requirements.txt
google-cloud-bigquery==2.26.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
pandas==1.3.3
numpy==1.21.2
PyYAML==5.4.1
